Better webdriver use
The whole idea of the webdriver control flow is that asynchronous tasks can be written in a asynchronous way, implicitly executing each promise when the one queued before is fullfiled. This style mix very well with cucumber.js and should be encouraged. All that is needed is to return the last promise at the end of a step.


cf: https://github.com/angular/protractor/blob/master/docs/control-flow.md